fix: migrate to KSP and resolve compilation errors
- Migrated Hilt from Kapt to KSP (Kotlin 2.0+ compatible)
- Fixed Int to Long type mismatches
- Removed duplicate composable declarations
- Fixed incorrect weight imports
- Added 50+ missing imports
- Fixed scope issues in LoginHeader
- Added togetherWith, scale, Color, Icons imports
- Reduced errors from 133 to 121 (9% improvement)

Still remaining: 121 errors (mostly state/logic issues)

+++ app/src/main/java/com/example/auratrackr/data/repository/UserRepositoryImpl.kt
@@ -163,7 +163,7 @@ class UserRepositoryImpl @Inject constructor(
                 .orderBy(FIELD_USERNAME)
                 .startAt(query)
                 .endAt(query + SEARCH_UNICODE_END)
-                .limit(SEARCH_LIMIT)
+                .limit(SEARCH_LIMIT.toLong())
 
             val snapshot = queryRef.get().await()
             snapshot.toObjects(User::class.java)
